feat: cleaned up componenet code, moved into compoenent folder, made map work awesome
