--- FeatureTests/Location.js
@@ -1,62 +1,0 @@
-'use strict';
-var React = require('react');
-var ReactNative = require('react-native');
-var { StyleSheet, Text, View, Button } = ReactNative;
-import { Icon } from 'native-base';
-
-exports.framework = 'React';
-exports.title = 'Geolocation';
-exports.description = 'Examples of using the Geolocation API.';
-exports.examples = [
-    {
-        title: 'navigator.geolocation',
-        render: function(): React.Element<any> {
-            return <GeolocationExample />;
-        },
-    }
-];
-
-import MapComp from './Map'
-import MapView from 'react-native-maps';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'white'
-  },
-  map: {
-    left: 0,
-    right: 0,
-    top: 0,
-    bottom: 0,
-    position: 'absolute'
-  },
-  marker: {
-    height: 20,
-    width: 20,
-    borderRadius: 20 / 2,
-    overflow: 'hidden'
-  },
-    title: {
-        fontWeight: '500',
-    },
-})
-export default class Location extends React.Component {
-  static navigationOptions = {
-    header: null,
-    tabBarIcon: ({ tintColor }) => (
-      <Icon ios='ios-map-outline' android="ios-map-outline" style={{color: tintColor}} />
-      )
-  }
-    constructor(props){
-        super(props)
-        this.state = {}
-    }
-    render(){
-        return (
-            <MapComp navigation={this.props.navigation}/>
-        )
-    }
-}
