+++ Nav.js
@@ -1,33 +1,23 @@
 import { StackNavigator, TabNavigator } from 'react-navigation';
-import React, { Component } from 'react';
-import PictureTest from './FeatureTests/PictureTest';
-import PostDetail from './FeatureTests/StyleTest';
-import VideoTest from './FeatureTests/VideoTest';
-import Location from './FeatureTests/Location';
-import App from './FeatureTests/src/App.js';
-import Login from './auth/Login.js';
-import User from './auth/SingleUser.js';
-import RenderVideoTest from './FeatureTests/RenderVideoTest';
-// import Post from './database/Post';
-// import SinglePost from './database/SinglePost';
-
+import React from 'react';
 import * as firebase from 'firebase';
 
+import VideoTest from './FeatureTests/VideoTest';
+import Map from './components/Map';
+import Login from './components/Login.js';
+import Account from './components/Account.js';
+import RenderVideoTest from './FeatureTests/RenderVideoTest';
+import App from './FeatureTests/src/App';
+
 const firebaseConfig = {
-    apiKey: "AIzaSyB8MNYp0Y5U6FztmjVWzILaPnYdKqntPN0",
-    authDomain: "we-ward-872a3.firebaseapp.com",
-    databaseURL: "https://we-ward-872a3.firebaseio.com",
-    projectId: "we-ward-872a3",
-    storageBucket: "we-ward-872a3.appspot.com",
-    messagingSenderId: "745916231980"
+    apiKey: 'AIzaSyB8MNYp0Y5U6FztmjVWzILaPnYdKqntPN0',
+    authDomain: 'we-ward-872a3.firebaseapp.com',
+    databaseURL: 'https://we-ward-872a3.firebaseio.com',
+    projectId: 'we-ward-872a3',
+    storageBucket: 'we-ward-872a3.appspot.com',
+    messagingSenderId: '745916231980'
 };
 export const firebaseApp = firebase.initializeApp(firebaseConfig);
-import {
-    AppRegistry,
-    StyleSheet,
-    Text,
-    View
-} from 'react-native';
 
 
 export default class Nav extends React.Component {
@@ -47,34 +37,25 @@ export default class Nav extends React.Component {
         firebaseApp.auth().onAuthStateChanged(user => {
             if (user) {
                 this.setState({
-                    user : user,
+                    user,
                     tab: TabNavigator({
                         View: {
-                            screen: Location
+                            screen: Map
                         },
-                        // PictureTest: {
-                        //     screen: PictureTest
+                        // App: {
+                        //     screen: App
                         // },
                         Share: {
                             screen: VideoTest
                         },
                         Account: {
-                            title: 'Account',
-                            screen: User
+                            screen: Account
                         },
-                        // Post: {
-                        //     screen: Post
-                        // },
-                        // SinglePost: {
-                        //     screen: SinglePost
-                        // }
                     })
-                })
+                });
             }
         });
     }
-
-
 
     render () {
         return (
@@ -91,9 +72,6 @@ export default class Nav extends React.Component {
                     }
                 }))
             )
-
-
-        )
+        );
     }
-
 }
