+++ components/GoogleSearch.js
@@ -1,44 +1,40 @@
-import React, {Component} from 'react';
-import {Text, Image} from 'react-native'
+import React, { Component } from 'react';
+import { Text, Icon } from 'native-base'
 
-var {GooglePlacesAutocomplete} = require('react-native-google-places-autocomplete');
+var { GooglePlacesAutocomplete } = require('react-native-google-places-autocomplete');
 
 const homePlace = {description: 'Home', geometry: { location: { lat: 48.8152937, lng: 2.4597668 } }};
 const workPlace = {description: 'Work', geometry: { location: { lat: 48.8496818, lng: 2.2940881 } }};
 const fullStack = {description: 'FullStack', geometry: { location: { lat: 40.7128, lng: -74.0093 } }};
 
-class GoogleSearch extends React.Component {
-  constructor(props) {
-    super(props)
-  }
+class GoogleSearch extends Component {
   render() {
     return (
       <GooglePlacesAutocomplete
-        placeholder='Search by location'
-        minLength={2} // minimum length of text to search
-        autoFocus={false}
-        listViewDisplayed='auto'    // true/false/undefined
+        placeholder="Search by location"
+        minLength={2}
+        autoFocus={true}
+        listViewDisplayed="auto"
         fetchDetails={true}
-        renderDescription={(row) => row.description} // custom description render
-        onPress={(data, details = null) => { // 'details' is provided when fetchDetails = true
-          this.props.onSearch(details.geometry.location)
+        renderDescription={row => row.description}
+        onPress={(data, details = null) => {
+          // 'details' is provided when fetchDetails = true
+          this.props.onSearch(details.geometry.location);
+          this.props.setModalVisible()
         }}
         getDefaultValue={() => {
-          return ''; // text input default value
+          // text input default value
+          return '';
         }}
         query={{
           // available options: https://developers.google.com/places/web-service/autocomplete
-          key: 'AIzaSyA1-0UvhC1C4ErCbeu01RriQSNrWPX3Wyw',
+          key: 'AIzaSyA1-0UvhC1C4ErCbeu01RriQSNrWPX3Wyw', //probably should hide this...
           language: 'en', // language of the results
-          types: '(cities)', // default: 'geocode'
+          types: 'geocode', // default: 'geocode'
         }}
         styles={{
-          container: {
-            flex: 1
-          },
           textInputContainer: {
             backgroundColor: 'rgba(0,0,0,0)',
-            // width: 350,
             borderTopWidth: 0,
             borderBottomWidth:0
           },
@@ -53,10 +49,9 @@ class GoogleSearch extends React.Component {
             color: '#1faadb'
           },
         }}
-
-        currentLocation={true} // Will add a 'Current location' button at the top of the predefined places list
+        currentLocation={true} // Not currently working... Will add a 'Current location' button at the top of the predefined places list
         currentLocationLabel="Current location"
-        nearbyPlacesAPI='GooglePlacesSearch' // Which API to use: GoogleReverseGeocoding or GooglePlacesSearch
+        nearbyPlacesAPI="GooglePlacesSearch" // Which API to use: GoogleReverseGeocoding or GooglePlacesSearch
         GoogleReverseGeocodingQuery={{
           // available options for GoogleReverseGeocoding API : https://developers.google.com/maps/documentation/geocoding/intro
         }}
@@ -65,15 +60,11 @@ class GoogleSearch extends React.Component {
           rankby: 'distance',
           types: 'food',
         }}
-
-
         filterReverseGeocodingByTypes={['locality', 'administrative_area_level_3']} // filter the reverse geocoding results by types - ['locality', 'administrative_area_level_3'] if you want to display only cities
-
-        predefinedPlaces={[homePlace, workPlace, fullStack]}
-
         debounce={200} // debounce the requests in ms. Set to 0 to remove debounce. By default 0ms.
-        // renderLeftButton={() => <Image source={require('path/custom/left-icon')} />}
-        renderRightButton={() => <Text></Text>}
+        predefinedPlaces={[fullStack]}
+        // renderLeftButton={() => <Icon name="ios-search-outline" />}
+        // renderRightButton={() => <Text>Cancel</Text>}
       />
     );
   }
